backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-jose[cryptography]>=3.3.0
passlib>=1.7.4
python-dotenv>=1.0.0
pydantic>=2.10.0
pydantic-settings>=2.6.0
openai>=1.57.0
anthropic>=0.40.0
google-generativeai>=0.8.3
httpx>=0.28.0
